Clean up magic components
